#22 Fix type inconsistencies in capital change DTOs and queries
Changed IncreasePercent property from double? to decimal? across
IncreaseCapitalDto, ExtraAssemblyIncreaseCapital domain entity, and
ExtraAssemblyIncreaseCapitalQueryDto for consistency with financial
data handling and DecreasePercent property.

Also made IsAccept properties nullable (bool?) in IncreaseCapitalDto
and ShareValueChangeCapitalDto to handle optional JSON fields and
prevent deserialization failures. Updated mapping service to use null
coalescing operator (??) with false default for null-to-non-null
conversions.

Build: Success (0 errors)
Tests: All 191 tests passed (36 domain, 18 buildingblock, 7 application,
3 performance, 127 integration)

+++ src/Application/Fundamental.Application/Codals/Manufacturing/Queries/GetExtraAssemblyById/GetExtraAssemblyDetailItem.cs
@@ -81,7 +81,7 @@ public sealed record ExtraAssemblyIncreaseCapitalQueryDto(
     int? SarfSaham,
     bool IsAccept,
     int? CapitalIncreaseValue,
-    double? IncreasePercent,
+    decimal? IncreasePercent,
     CapitalIncreaseApprovalType Type,
     decimal? CashForceclosurePriorityStockPrice,
     string? CashForceclosurePriorityStockDesc,
